+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/part/explorer/project/ProjectExplorerPresenter.java
@@ -478,26 +478,30 @@ public class ProjectExplorerPresenter extends BasePresenter
 
     @Override
     public void onExecute() {
-      if (view.getTree().getNodeLoader().isBusy()) {
-        delay(500);
+      Scheduler.get()
+          .scheduleDeferred(
+              () -> {
+                if (view.getTree().getNodeLoader().isBusy()) {
+                  delay(500);
 
-        return;
-      }
+                  return;
+                }
 
-      final Set<Path> updateQueue = Sets.newHashSet(toRefresh);
-      toRefresh.clear();
+                final Set<Path> updateQueue = Sets.newHashSet(toRefresh);
+                toRefresh.clear();
 
-      for (Path path : updateQueue) {
-        final Node node = getNode(path);
+                for (Path path : updateQueue) {
+                  final Node node = getNode(path);
 
-        if (node == null) {
-          continue;
-        }
+                  if (node == null) {
+                    continue;
+                  }
 
-        if (getTree().isExpanded(node)) {
-          view.getTree().getNodeLoader().loadChildren(node, true);
-        }
-      }
+                  if (getTree().isExpanded(node)) {
+                    view.getTree().getNodeLoader().loadChildren(node, true);
+                  }
+                }
+              });
     }
   }
 }
