--- wsagent/che-core-api-project/src/main/java/org/eclipse/che/api/project/server/ProjectProblemImpl.java
@@ -1,25 +1,0 @@
-package org.eclipse.che.api.project.server;
-
-import org.eclipse.che.api.core.model.project.ProjectProblem;
-
-/** @author Vitalii Parfonov */
-public class ProjectProblemImpl implements ProjectProblem {
-
-  public ProjectProblemImpl(int code, String message) {
-    this.code = code;
-    this.message = message;
-  }
-
-  int code;
-  String message;
-
-  @Override
-  public int getCode() {
-    return code;
-  }
-
-  @Override
-  public String getMessage() {
-    return message;
-  }
-}
