+++ wsagent/che-core-api-project/src/main/java/org/eclipse/che/api/project/server/ProjectTypes.java
@@ -26,6 +26,7 @@ import org.eclipse.che.api.core.model.project.ProjectProblem;
 import org.eclipse.che.api.core.model.project.type.Attribute;
 import org.eclipse.che.api.project.server.type.ProjectTypeDef;
 import org.eclipse.che.api.project.server.type.ProjectTypeRegistry;
+import org.eclipse.che.api.workspace.server.model.impl.ProjectProblemImpl;
 
 /** @author gazarenkov */
 public class ProjectTypes {
